+++ R/chisq-methods.R
@@ -18,6 +18,11 @@ setMethod("chisq", "SummaryEL", function(object, ...) {
   object@statistic
 })
 
+#' @describeIn chisq Extracts the chi-square statistic.
+setMethod("chisq", "SummaryELT", function(object, ...) {
+  object@statistic
+})
+
 #' @describeIn chisq Extracts the chi-square statistic for the overall test of
 #'   the model.
 setMethod("chisq", "SummaryLM", function(object, ...) {