+++ R/getOptim-methods.R
@@ -7,3 +7,18 @@ setMethod("getOptim", "EL", function(object, ...) {
 setMethod("getOptim", "ELT", function(object, ...) {
   object@optim
 })
+
+#' @rdname getOptim
+setMethod("getOptim", "SummaryEL", function(object, ...) {
+  object@optim
+})
+
+#' @rdname getOptim
+setMethod("getOptim", "SummaryELT", function(object, ...) {
+  object@optim
+})
+
+#' @rdname getOptim
+setMethod("getOptim", "SummaryLM", function(object, ...) {
+  object@optim
+})