+++ R/pVal-methods.R
@@ -19,6 +19,11 @@ setMethod("pVal", "SummaryEL", function(object, ...) {
 })
 
 #' @describeIn pVal Extracts the \eqn{p}-value.
+setMethod("pVal", "SummaryELT", function(object, ...) {
+  object@pval
+})
+
+#' @describeIn pVal Extracts the \eqn{p}-value.
 setMethod("pVal", "SummaryLM", function(object, ...) {
   object@pval
 })